+++ app/trigger/internal/logic/listplanslogic.go
@@ -49,9 +49,9 @@ func (l *ListPlansLogic) ListPlans(in *trigger.ListPlansReq) (*trigger.ListPlans
 		db = db.Where("type = ?", in.Type)
 	}
 	if len(in.Status) > 0 {
-		statusInts := make([]int64, len(in.Status))
+		statusInts := make([]int, len(in.Status))
 		for i, status := range in.Status {
-			statusInts[i] = int64(status)
+			statusInts[i] = int(status)
 		}
 		db = db.Where("status IN ?", statusInts)
 	}
