+++ app/trigger/internal/logic/pauseplanexecitemlogic.go
@@ -66,26 +66,26 @@ func (l *PausePlanExecItemLogic) PausePlanExecItem(in *trigger.PausePlanExecItem
 	if err := db.Where("plan_id = ?", execItem.PlanId).First(&plan).Error; err != nil {
 		return nil, tool.NewErrorByPbCodeWrap(extproto.Code__1_02_DB, err, "查询计划失败")
 	}
-	if plan.Status == int64(model.PlanStatusTerminated) || plan.FinishedTime.Valid {
+	if plan.Status == model.PlanStatusTerminated || plan.FinishedTime.Valid {
 		return nil, tool.NewErrorByPbCode(extproto.Code__1_05_BIZ_STATE, "计划状态已结束,无需暂停")
 	}
 
-	if planBatch.Status == int64(model.PlanStatusTerminated) || planBatch.FinishedTime.Valid {
+	if planBatch.Status == model.PlanStatusTerminated || planBatch.FinishedTime.Valid {
 		return nil, tool.NewErrorByPbCode(extproto.Code__1_05_BIZ_STATE, "计划批次状态已结束,无需暂停")
 	}
 
-	if execItem.Status == int64(model.StatusCompleted) || execItem.Status == int64(model.StatusTerminated) || execItem.Status == int64(model.StatusPaused) {
+	if execItem.Status == model.StatusCompleted || execItem.Status == model.StatusTerminated || execItem.Status == model.StatusPaused {
 		return nil, tool.NewErrorByPbCode(extproto.Code__1_05_BIZ_STATE, "执行项状态已结束,无需暂停")
 	}
 
-	if execItem.Status == int64(model.StatusRunning) {
+	if execItem.Status == model.StatusRunning {
 		return &trigger.PausePlanExecItemRes{}, tool.NewErrorByPbCode(extproto.Code__1_05_BIZ_STATE, "执行项正在运行中，请稍后再试")
 	}
 
 	// 执行事务
 	err = db.Transaction(func(tx *gorm.DB) error {
 		// 更新执行项状态为暂停
-		execItem.Status = int64(model.StatusPaused)
+		execItem.Status = model.StatusPaused
 		execItem.PausedTime = sql.NullTime{Time: time.Now(), Valid: true}
 		execItem.PausedReason = sql.NullString{String: in.Reason, Valid: in.Reason != ""}
 		execItem.UpdateUser = sql.NullString{String: tool.GetCurrentUserId(l.ctx, nil), Valid: tool.GetCurrentUserId(l.ctx, nil) != ""}
