+++ app/trigger/internal/logic/runplanexecitemlogic.go
@@ -57,7 +57,7 @@ func (l *RunPlanExecItemLogic) RunPlanExecItem(in *trigger.RunPlanExecItemReq) (
 		}
 		return nil, tool.NewErrorByPbCode(extproto.Code__1_02_DB, "查询执行项失败")
 	}
-	if execItem.Status != int64(model.StatusWaiting) && execItem.Status != int64(model.StatusDelayed) {
+	if execItem.Status != model.StatusWaiting && execItem.Status != model.StatusDelayed {
 		return nil, tool.NewErrorByPbCode(extproto.Code__1_05_BIZ_STATE, fmt.Sprintf("执行项当前状态为%d，无法立即执行，仅支持等待调度(0)或延期等待(10)状态", execItem.Status))
 	}
 	// 查询计划批次
@@ -72,19 +72,19 @@ func (l *RunPlanExecItemLogic) RunPlanExecItem(in *trigger.RunPlanExecItemReq) (
 		return nil, tool.NewErrorByPbCodeWrap(extproto.Code__1_02_DB, err, "查询计划失败")
 	}
 
-	if plan.Status == int64(model.PlanStatusTerminated) || plan.FinishedTime.Valid {
+	if plan.Status == model.PlanStatusTerminated || plan.FinishedTime.Valid {
 		return nil, tool.NewErrorByPbCode(extproto.Code__1_05_BIZ_STATE, "计划状态已结束,不可立即执行")
 	}
 
-	if planBatch.Status == int64(model.PlanStatusTerminated) || planBatch.FinishedTime.Valid {
+	if planBatch.Status == model.PlanStatusTerminated || planBatch.FinishedTime.Valid {
 		return nil, tool.NewErrorByPbCode(extproto.Code__1_05_BIZ_STATE, "计划批次状态已结束,不可立即执行")
 	}
 
-	if plan.Status == int64(model.PlanStatusPaused) {
+	if plan.Status == model.PlanStatusPaused {
 		return nil, tool.NewErrorByPbCode(extproto.Code__1_05_BIZ_STATE, "计划处于暂停状态,不可立即执行")
 	}
 
-	if planBatch.Status == int64(model.PlanStatusPaused) {
+	if planBatch.Status == model.PlanStatusPaused {
 		return nil, tool.NewErrorByPbCode(extproto.Code__1_05_BIZ_STATE, "计划批次处于暂停状态,不可立即执行")
 	}
